backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
PyJWT==2.8.0
django-environ==0.11.2
dj-database-url==2.1.0
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2
face-recognition==1.3.0
dlib==19.24.2
opencv-python==4.8.1.78
twilio==8.10.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
locust==2.17.0
sentry-sdk==1.38.0
